+++ src/scripts/module.js
@@ -77,7 +77,7 @@
 
         $scope.processing = false;
         $scope.mode = mode;
-        $scope.entity = scopeEntityFromEntity(formScheme, entity);
+        $scope.entity = scopeEntityFromEntity(entity);
         $scope.formScheme = formScheme;
 
         $scope.save = function (form) {
@@ -110,7 +110,7 @@
         };
 
         $scope.reset = function () {
-          $scope.entity = scopeEntityFromEntity(formScheme, entity);
+          $scope.entity = scopeEntityFromEntity(entity);
         };
 
         $scope.cancel = function () {
@@ -137,7 +137,7 @@
           } else {
             return url;
           }
-        }
+        };
 
       }
 
@@ -177,22 +177,12 @@
    * Returns entity to be used inside of scope,
    * based on entity received from the repository.
    *
-   * @param {object} formScheme
    * @param {object} entity
    *
    * @returns {object}
    */
-  function scopeEntityFromEntity (formScheme, entity) {
-    var result = {};
-    iterateFields(formScheme, function (fieldName) {
-      var value = eval('entity.' + fieldName);
-      // Casting moments to vanilla dates.
-      //if (moment.isMoment(value)) {
-      //  value = value.toDate();
-      //}
-      propertyPathSet(result, fieldName, value);
-    });
-    return result;
+  function scopeEntityFromEntity (entity) {
+    return angular.extend({}, entity);
   }
 
   /**
@@ -207,8 +197,8 @@
   function getDataToSave (formScheme, scopeEntity) {
     var result = {};
     iterateFields(formScheme, function (fieldName, field) {
-      if (scopeEntity[fieldName] && !field.readonly) {
-        result[fieldName] = (scopeEntity[fieldName]);
+      if (!field.readonly && !field.hidden) {
+        propertyPathSet(result, fieldName, propertyPathGet(scopeEntity, fieldName));
       }
     });
     return result;
@@ -243,6 +233,17 @@
         case 'datetime':
         case 'date':
           field.elementType = field.type;
+          break;
+        case 'enum':
+          if (!field.selectExpression) {
+            field.selectExpression = 'key as value for (key, value)';
+          }
+          break;
+        case 'text':
+          if (!field.rowsCount) {
+            field.rowsCount = 10;
+          }
+          break;
       }
       switch (field.type) {
         case 'string':
@@ -265,7 +266,7 @@
    * Iterates specified form scheme fields and
    * executes a specified callback function for each field.
    *
-   * @param {array} formScheme
+   * @param {object} formScheme
    * @param {function} cb
    */
   function iterateFields (formScheme, cb) {
@@ -332,6 +333,13 @@
     }
   }
 
+  /**
+   * Sets object property value by it's path.
+   *
+   * @param {Object} object
+   * @param {string} path
+   * @param {*} value
+   */
   function propertyPathSet (object, path, value) {
     var parts = path.split('.');
     var reference = object;
@@ -347,6 +355,18 @@
       i++;
     });
     reference = value;
+  }
+
+  /**
+   * Returns property value from object by it's path.
+   *
+   * @param {Object} object
+   * @param {string} path
+   *
+   * @returns {Object}
+   */
+  function propertyPathGet (object, path) {
+    return eval('object.' + path);
   }
 
 })(window, angular);